refactor: collection tabs routing
