+++ src/components/Spinner/Spinner.tsx
@@ -10,7 +10,9 @@ export const Spinner: FC = () => {
       style={{
         display: 'flex',
         justifyContent: 'center',
+        alignItems: 'center',
         width: '100%',
+        height: '100%',
       }}
     >
       <Spin size="large" indicator={antIcon} />
