+++ src/constants/paths.ts
@@ -1,14 +1,18 @@
-export enum COLLECTION_TABS {
-  BUY = 'buy',
-  SELL = 'sell',
-  ACTIVITY = 'activity',
-  POOLS = 'pools',
-}
+// export enum COLLECTION_TABS {
+//   BUY = 'buy',
+//   SELL = 'sell',
+//   ACTIVITY = 'activity',
+//   POOLS = 'pools',
+// }
 
+// export const createCollectionLink = (
+//   tab = COLLECTION_TABS.BUY,
+//   publicKey = ':publicKey',
+// ): string => `/collection/${publicKey}/${tab}`;
 export const createCollectionLink = (
-  tab = COLLECTION_TABS.BUY,
+  //tab = COLLECTION_TABS.BUY,
   publicKey = ':publicKey',
-): string => `/collection/${publicKey}/${tab}`;
+): string => `/collection/${publicKey}`;
 
 export const createCreatePollLink = (
   publicKey = ':publicKey?',
@@ -25,11 +29,11 @@ export const createEditPollLink = (poolPubkey = ':poolPubkey?'): string =>
 export const PATHS = {
   ROOT: '/', //? Main page with common bullshit texts
   COLLECTIONS: '/collections', //? Collections table|list
-  COLLECTION: '/collection', //? Collections table|list
-  COLLECTION_BUY: createCollectionLink(COLLECTION_TABS.BUY), //? Specific collection buy tab. Should be default tab if tab isn't specified
-  COLLECTION_SELL: createCollectionLink(COLLECTION_TABS.SELL), //? Specific collection sell tab.
-  COLLECTION_ACTIVITY: createCollectionLink(COLLECTION_TABS.ACTIVITY), //? Specific collection activity tab.
-  COLLECTION_POOLS: createCollectionLink(COLLECTION_TABS.POOLS), //? Specific collection pools tab.
+  COLLECTION: createCollectionLink(), //? Collections table|list
+  // COLLECTION_BUY: createCollectionLink(COLLECTION_TABS.BUY), //? Specific collection buy tab. Should be default tab if tab isn't specified
+  // COLLECTION_SELL: createCollectionLink(COLLECTION_TABS.SELL), //? Specific collection sell tab.
+  // COLLECTION_ACTIVITY: createCollectionLink(COLLECTION_TABS.ACTIVITY), //? Specific collection activity tab.
+  // COLLECTION_POOLS: createCollectionLink(COLLECTION_TABS.POOLS), //? Specific collection pools tab.
   POOL_PAGE: '/pools/:poolPubkey', //? Specific pool page.
   POOL_EDIT: createEditPollLink(), //? Page to create pool.
   CREATE_POOL: createCreatePollLink(), //? Page to create pool.
