+++ src/constants/routes.ts
@@ -4,15 +4,12 @@ import { PATHS } from './paths';
 import { Home } from '../pages/HomePage/Home';
 import { Page404 } from '../pages/Page404/Page404';
 import { Collections } from '../pages/Collections/Collections';
-import { CollectionActivityPage } from '../pages/Collection/CollectionActivityPage';
+import { CollectionPage } from '../pages/Collection/CollectionPage';
 import { MyNfts } from '../pages/MyNfts/MyNfts';
 import { MyPools } from '../pages/MyPools/MyPools';
-import { CollectionPoolsPage } from '../pages/Collection/CollectionPoolsPage';
 import { PoolPage } from '../pages/Pool/PoolPage';
 import { NewPool } from '../pages/NewPool/NewPool';
 import { EditPool } from '../pages/EditPool/EditPool';
-import { CollectionBuyPage } from '../pages/Collection/CollectionBuyPage';
-import { CollectionSellPage } from '../pages/Collection/CollectionSellPage';
 
 interface Route {
   path: string;
@@ -33,23 +30,8 @@ export const routes: Route[] = [
   },
   {
     exact: true,
-    path: PATHS.COLLECTION_BUY,
-    component: CollectionBuyPage,
-  },
-  {
-    exact: true,
-    path: PATHS.COLLECTION_SELL,
-    component: CollectionSellPage,
-  },
-  {
-    exact: true,
-    path: PATHS.COLLECTION_ACTIVITY,
-    component: CollectionActivityPage,
-  },
-  {
-    exact: true,
-    path: PATHS.COLLECTION_POOLS,
-    component: CollectionPoolsPage,
+    path: PATHS.COLLECTION,
+    component: CollectionPage,
   },
   {
     exact: true,
