+++ src/pages/Collection/components/CollectionInfo/CollectionInfo.constants.tsx
