+++ src/pages/Collection/components/CollectionTabs/styles.module.scss
@@ -1,6 +1,5 @@
 .collectionTabsWrapper {
   margin-top: 16px;
-  margin-bottom: 24px;
 
   :global {
     .ant-tabs-nav {
@@ -81,4 +80,100 @@
   align-items: center;
   justify-content: center;
   min-width: 38px;
+}
+
+.tabContentWrapper {
+  margin-top: 24px;
+}
+
+.cards {
+  display: grid;
+  grid-template-columns: repeat(auto-fill, minmax(158px, 1fr));
+  gap: 16px;
+  align-items: stretch;
+
+  @media screen and (max-width: 767px) {
+    grid-template-columns: repeat(auto-fill, minmax(120px, 1fr));
+  }
+}
+
+// TODO: refactor this
+.activityCards {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  align-items: stretch;
+  gap: 16px;
+
+  @media screen and (max-width: 1024px) {
+    gap: 8px;
+  }
+}
+
+.activityCard {
+  display: flex;
+  align-items: center;
+  background: #272f37;
+  border-radius: 20px;
+  font-size: 16px;
+  min-height: 64px;
+
+  @media screen and (max-width: 1024px) {
+    flex-direction: column;
+    padding: 12px 8px;
+  }
+}
+
+.activityLink {
+  color: #9dcaff;
+  text-decoration: underline;
+
+  &:hover,
+  &:active {
+    color: #d1e4ff;
+    text-decoration: underline;
+  }
+}
+
+.activityTaker {
+  margin-bottom: 0;
+  margin-right: 8px;
+
+  @media screen and (max-width: 1024px) {
+    margin-right: 0;
+    width: 100%;
+    text-align: center;
+  }
+}
+
+.activityNftImage {
+  width: 64px;
+  border-radius: 20px;
+
+  @media screen and (max-width: 1024px) {
+    margin-bottom: 8px;
+  }
+}
+
+.activityNftName {
+  margin: 0 8px;
+}
+
+.activityPrice {
+  display: flex;
+  margin-bottom: 0;
+}
+
+.activityPool {
+  margin-bottom: 0;
+  margin-right: 8px;
+
+  @media screen and (max-width: 1024px) {
+    margin-right: 0;
+  }
+}
+
+.activityDate {
+  display: inline-block;
+  margin-bottom: 0;
 }
