+++ src/pages/Collection/components/MakeOfferModal/styles.module.scss
@@ -1,0 +1,6 @@
+.nftCard {
+  display: flex;
+  flex-direction: row;
+  column-gap: 20px;
+  margin-bottom: 20px;
+}
